refac(track/Handler): AddBranch => TrackBranch
Renames AddBranch to TrackBranch to better reflect its purpose.
